Refactor Engine Variables and Package Imports
