+++ src/pyorlib/mp/algebra/__init__.py
@@ -1,1 +1,13 @@
 from .element import Element
+from .expressions import Expression
+from .terms import Term
+from .terms.constants import Constant
+from .terms.variables import Variable
+
+__all__ = [
+    "Element",
+    "Expression",
+    "Term",
+    "Constant",
+    "Variable",
+]
