--- src/pyorlib/mp/algebra/terms/variables/cplex/__init__.py
@@ -1,1 +1,0 @@
-from .cplex_variable import CplexVariable
