--- src/pyorlib/mp/algebra/terms/variables/gurobi/__init__.py
@@ -1,1 +1,0 @@
-from .gurobi_variable import GurobiVariable
