--- src/pyorlib/mp/algebra/terms/variables/ortools/__init__.py
@@ -1,1 +1,0 @@
-from .ortools_variable import ORToolsVariable
