--- src/pyorlib/mp/algebra/terms/variables/pulp/__init__.py
@@ -1,1 +1,0 @@
-from .pulp_variable import PuLPVariable
