+++ src/pyorlib/mp/structures/__init__.py
@@ -1,0 +1,11 @@
+from .definitions import DimensionDefinition, ParameterDefinition, TermDefinition
+from .parameters import MultiValueParameter, Parameter, SingleValueParameter
+
+__all__ = [
+    "DimensionDefinition",
+    "ParameterDefinition",
+    "TermDefinition",
+    "MultiValueParameter",
+    "Parameter",
+    "SingleValueParameter",
+]
