+++ tests/mp/validators/test_value_type_validator.py
@@ -1,6 +1,6 @@
 from math import inf
 
-from pyorlib import ValueTypeValidator
+from pyorlib.mp.validators import ValueTypeValidator
 
 
 class TestValueTypeValidator:
